requirements.txt
google-cloud-storage
google-cloud-pubsub
board
adafruit-circuitpython-dht
picamera
docker
